Improve consistency in docs and module structure
